Rename files to avoid case sensitivity issues
