+++ functions/compare-loose.js
