+++ functions/compare.js
@@ -1,4 +1,4 @@
-var SemVer = require('../SemVer');
+var SemVer = require('../semver');
 
 module.exports = function compare (a, b, loose) {
   return new SemVer(a, loose).compare(new SemVer(b, loose))
