+++ functions/major.js
@@ -1,4 +1,4 @@
-var SemVer = require('../SemVer');
+var SemVer = require('../semver');
 
 module.exports = function major (a, loose) {
   return new SemVer(a, loose).major
