+++ functions/rsort.js
@@ -1,4 +1,4 @@
-var compareBuild = require('./compareBuild');
+var compareBuild = require('./compare-build');
 
 module.exports = function rsort (list, loose) {
     return list.sort(function (a, b) {
