+++ functions/sort.js
@@ -1,4 +1,4 @@
-var compareBuild = require('./compareBuild');
+var compareBuild = require('./compare-build');
 
 module.exports = function sort (list, loose) {
     return list.sort(function (a, b) {
