+++ index.js
@@ -1,4 +1,4 @@
-var SemVer = require('./SemVer');
+var SemVer = require('./semver');
 var { compareIdentifiers, rcompareIdentifiers } = require('./internal/identifiers');
 var { SEMVER_SPEC_VERSION } = require('./internal/constants');
 var { re, src, t } = require('./internal/re');
@@ -24,8 +24,8 @@ exports.patch = patch = require('./functions/patch');
 exports.prerelease = prerelease = require('./functions/prerelease');
 exports.compare = compare = require('./functions/compare');
 exports.rcompare = rcompare = require('./functions/rcompare');
-exports.compareLoose = compareLoose = require('./functions/compareLoose');
-exports.compareBuild = compareBuild = require('./functions/compareBuild');
+exports.compareLoose = compareLoose = require('./functions/compare-loose');
+exports.compareBuild = compareBuild = require('./functions/compare-build');
 exports.sort = sort = require('./functions/sort');
 exports.rsort = rsort = require('./functions/rsort');
 exports.gt = gt = require('./functions/gt');
@@ -42,11 +42,11 @@ exports.Comparator = Comparator;
 exports.Range = Range;
 exports.satisfies = satisfies;
 
-exports.toComparators = toComparators = require('./ranges/toComparators');
-exports.maxSatisfying = maxSatisfying = require('./ranges/maxSatisfying');
-exports.minSatisfying = minSatisfying = require('./ranges/minSatisfying');
-exports.minVersion = minVersion = require('./ranges/minVersion');
-exports.validRange = validRange = require('./ranges/validRange');
+exports.toComparators = toComparators = require('./ranges/to-comparators');
+exports.maxSatisfying = maxSatisfying = require('./ranges/max-satisfying');
+exports.minSatisfying = minSatisfying = require('./ranges/min-satisfying');
+exports.minVersion = minVersion = require('./ranges/min-version');
+exports.validRange = validRange = require('./ranges/valid-range');
 exports.outside = outside = require('./ranges/outside');
 exports.gtr = gtr = require('./ranges/gtr');
 exports.ltr = ltr = require('./ranges/ltr');
