+++ ranges/to-comparators.js
